Make sure we bind as the configured user.
In some cases, we are bound as the current user and if that
user doesn't have enough rights on the LDAP server the search
will fail. Be sure we bind with the user that Horde was configured
to bind with.

Fixes failing to find user calendars during CalDAV access with
using LDAP authentication and restrictive permissions on the LDAP
server.